Introduced “GrailsTestTypeRunner” and “GrailsTestTypeResult” as the new contract for test types to be run by `grails test-app`.
This layer of abstraction is necessary to completely separate the test running infrastructure from what actual runs the tests. This change makes it possible to easily plug in a TestNG runner for example.

Note: This is WIP. This is the first phase which introduced the new API and adapted the old one to work with it.