Introduced “GrailsTestEventPublisher”.
An instance of this class is passed to each test runner. Test runners are responsible for calling it's methods in the correct sequence.

This is really about facilitating IDE integration for running tests. IDEs can use the Grails “build listening” feature to inject a listener for these events and display suitable visual notification of test run results.